QuantumESPRESSOCommands

Documentation for QuantumESPRESSOCommands.

QuantumESPRESSOCommands.COMMAND_ENTRY_DOC_STUB

QuantumESPRESSOCommands.comonicon_install

QuantumESPRESSOCommands.comonicon_install_path

QuantumESPRESSOCommands.COMMAND_ENTRY_DOC_STUB — Constant

The main command qe.

QuantumESPRESSOCommands.comonicon_install — Method

comonicon_install(;kwargs...)

Install the CLI manually. This will use the default configuration in Comonicon.toml, if it exists. See also comonicon_build. For more detailed reference, please refer to Comonicon documentation.

QuantumESPRESSOCommands.comonicon_install_path — Method

comonicon_install_path([quiet=false])

Install the PATH and FPATH to your shell configuration file. You can use comonicon_install_path(true), to skip interactive prompt. For more detailed reference, please refer to Comonicon documentation.