API

Public interfaces

QuantumESPRESSOCommands.pw — Function

Run command pw.x.

Arguments

input: the path to the input file.

output: the path to the output file. If not specified, use a temporary path.

Options

--np <n>: the number of processes used.

--path <path>: the path to the executable.

Flags

--chdir: if true, change directory to where the input file is stored when running.

QuantumESPRESSOCommands.ph — Function

Run command ph.x.

Arguments

input: the path to the input file.

output: the path to the output file. If not specified, use a temporary path.

Options

--np <n>: the number of processes used.

--path <path>: the path to the executable.

Flags

--chdir: if true, change directory to where the input file is stored when running.

QuantumESPRESSOCommands.q2r — Function

Run command q2r.x.

Arguments

input: the path to the input file.

output: the path to the output file. If not specified, use a temporary path.

Options

--np <n>: the number of processes used.

--path <path>: the path to the executable.

Flags

--chdir: if true, change directory to where the input file is stored when running.

QuantumESPRESSOCommands.matdyn — Function

Run command matdyn.x.

Arguments

input: the path to the input file.

output: the path to the output file. If not specified, use a temporary path.

Options

--np <n>: the number of processes used.

--path <path>: the path to the executable.

Flags

--chdir: if true, change directory to where the input file is stored when running.

QuantumESPRESSOCommands.dynmat — Function

Run command dynmat.x.

Arguments

input: the path to the input file.

output: the path to the output file. If not specified, use a temporary path.

Options

--np <n>: the number of processes used.

--path <path>: the path to the executable.

Flags

--chdir: if true, change directory to where the input file is stored when running.

Private interfaces

Missing docstring.

Missing docstring for ParallelizationFlags. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PwxConfig. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PhxConfig. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Q2rxConfig. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MatdynxConfig. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DynmatxConfig. Check Documenter's build log for details.

Missing docstring.

Missing docstring for makecmd. Check Documenter's build log for details.