API

Public interfaces

QuantumESPRESSOCommands.pw — Function

Run command pw.x.

Arguments

input: the path to the input file.

output: the path to the output file. If not specified, use a temporary path.

Options

--np <n>: the number of processes used.

--path <path>: the path to the executable.

Flags

--chdir: if true, change directory to where the input file is stored when running.

QuantumESPRESSOCommands.ph — Function

Run command ph.x.

Arguments

input: the path to the input file.

output: the path to the output file. If not specified, use a temporary path.

Options

--np <n>: the number of processes used.

--path <path>: the path to the executable.

Flags

--chdir: if true, change directory to where the input file is stored when running.

QuantumESPRESSOCommands.q2r — Function

Run command q2r.x.

Arguments

input: the path to the input file.

output: the path to the output file. If not specified, use a temporary path.

Options

--np <n>: the number of processes used.

--path <path>: the path to the executable.

Flags

--chdir: if true, change directory to where the input file is stored when running.

QuantumESPRESSOCommands.matdyn — Function

Run command matdyn.x.

Arguments

input: the path to the input file.

output: the path to the output file. If not specified, use a temporary path.

Options

--np <n>: the number of processes used.

--path <path>: the path to the executable.

Flags

--chdir: if true, change directory to where the input file is stored when running.

QuantumESPRESSOCommands.dynmat — Function

Run command dynmat.x.

Arguments

input: the path to the input file.

output: the path to the output file. If not specified, use a temporary path.

Options

--np <n>: the number of processes used.

--path <path>: the path to the executable.

Flags

--chdir: if true, change directory to where the input file is stored when running.

Private interfaces

QuantumESPRESSOCommands.ParallelizationFlags — Type

ParallelizationFlags(; nimage=0, npool=0, ntg=0, nyfft=0, nband=0, ndiag=0)

Construct parallelization flags of QuantumESPRESSO commands.

QuantumESPRESSOCommands.PwxConfig — Type

PwxConfig(; path, chdir, options)

Create configurations for pw.x.

Arguments

path::String="pw.x": the path to the executable.

chdir::Bool=true: whether to change directory to where the input file is stored when running pw.x. If false, stay in the current directory.

options::ParallelizationFlags=ParallelizationFlags(): the parallelization flags of pw.x.

QuantumESPRESSOCommands.PhxConfig — Type

PhxConfig(; path, chdir, options)

Create configurations for ph.x.

Arguments

path::String="ph.x": the path to the executable.

chdir::Bool=true: whether to change directory to where the input file is stored when running ph.x. If false, stay in the current directory.

options::ParallelizationFlags=ParallelizationFlags(): the parallelization flags of ph.x.

QuantumESPRESSOCommands.Q2rxConfig — Type

Q2rxConfig(; path, chdir, options)

Create configurations for q2r.x.

Arguments

path::String="q2r.x": the path to the executable.

chdir::Bool=true: whether to change directory to where the input file is stored when running q2r.x. If false, stay in the current directory.

options::ParallelizationFlags=ParallelizationFlags(): the parallelization flags of q2r.x.

QuantumESPRESSOCommands.MatdynxConfig — Type

MatdynxConfig(; path, chdir, options)

Create configurations for matdyn.x.

Arguments

path::String="matdyn.x": the path to the executable.

chdir::Bool=true: whether to change directory to where the input file is stored when running matdyn.x. If false, stay in the current directory.

options::ParallelizationFlags=ParallelizationFlags(): the parallelization flags of matdyn.x.

QuantumESPRESSOCommands.DynmatxConfig — Type

DynmatxConfig(; path, chdir, options)

Create configurations for dynmat.x.

Arguments

path::String="dynmat.x": the path to the executable.

chdir::Bool=true: whether to change directory to where the input file is stored when running dynmat.x. If false, stay in the current directory.

options::ParallelizationFlags=ParallelizationFlags(): the parallelization flags of dynmat.x.

QuantumESPRESSOCommands.makecmd — Function

makecmd(input, output; dir, mpi, main)

Make commands for QuantumESPRESSO executables.

Arguments

input: the path to the input file.

output=mktemp(parentdir(input))[1]: the path to the output file.

mpi=MpiexecConfig(): MPI configurations.

main: the configurations of the main executable.