API

Public interfaces

QuantumESPRESSOCommands.pw — Function

pw(input, output, error; use_script, mpi, main, cfgfile)

Run command pw.x.

Arguments

input: the path to the input file.

output=mktemp(parentdir(input))[1]: the path to the output file.

error=output: the path to the error file. By default, it logs into the output file.

use_script=false: if true, generate a shell script (with a random name) under the directory where the input file is stored, and run it.

mpi=MpiexecConfig(): MPI configurations.

main=PwxConfig(): the configurations of the main executable. In this case, pw.x.

cfgfile="": if not empty, load these configurations from a file.

QuantumESPRESSOCommands.ph — Function

ph(input, output, error; use_script, mpi, main, cfgfile)

Run command ph.x.

Arguments

input: the path to the input file.

output=mktemp(parentdir(input))[1]: the path to the output file.

error=output: the path to the error file. By default, it logs into the output file.

use_script=false: if true, generate a shell script (with a random name) under the directory where the input file is stored, and run it.

mpi=MpiexecConfig(): MPI configurations.

main=PhxConfig(): the configurations of the main executable. In this case, ph.x.

cfgfile="": if not empty, load these configurations from a file.

QuantumESPRESSOCommands.q2r — Function

q2r(input, output, error; use_script, mpi, main, cfgfile)

Run command q2r.x.

Arguments

input: the path to the input file.

output=mktemp(parentdir(input))[1]: the path to the output file.

error=output: the path to the error file. By default, it logs into the output file.

use_script=false: if true, generate a shell script (with a random name) under the directory where the input file is stored, and run it.

mpi=MpiexecConfig(): MPI configurations.

main=Q2rxConfig(): the configurations of the main executable. In this case, q2r.x.

cfgfile="": if not empty, load these configurations from a file.

QuantumESPRESSOCommands.matdyn — Function

matdyn(input, output, error; use_script, mpi, main, cfgfile)

Run command matdyn.x.

Arguments

input: the path to the input file.

output=mktemp(parentdir(input))[1]: the path to the output file.

error=output: the path to the error file. By default, it logs into the output file.

use_script=false: if true, generate a shell script (with a random name) under the directory where the input file is stored, and run it.

mpi=MpiexecConfig(): MPI configurations.

main=MatdynxConfig(): the configurations of the main executable. In this case, matdyn.x.

cfgfile="": if not empty, load these configurations from a file.

QuantumESPRESSOCommands.dynmat — Function

dynmat(input, output, error; use_script, mpi, main, cfgfile)

Run command dynmat.x.

Arguments

input: the path to the input file.

output=mktemp(parentdir(input))[1]: the path to the output file.

error=output: the path to the error file. By default, it logs into the output file.

use_script=false: if true, generate a shell script (with a random name) under the directory where the input file is stored, and run it.

mpi=MpiexecConfig(): MPI configurations.

main=DynmatxConfig(): the configurations of the main executable. In this case, dynmat.x.

cfgfile="": if not empty, load these configurations from a file.

Private interfaces

QuantumESPRESSOCommands.ParallelizationFlags — Type

ParallelizationFlags(; nimage=0, npool=0, ntg=0, nyfft=0, nband=0, ndiag=0)

Construct parallelization flags of QuantumESPRESSO commands.

QuantumESPRESSOCommands.PwxConfig — Type

PwxConfig(; exe, chdir, options)

Create configurations for pw.x.

Arguments

exe::String="pw.x": the path to the executable.

chdir::Bool=true: whether to change directory to where the input file is stored when running pw.x. If false, stay in the current directory.

options::ParallelizationFlags=ParallelizationFlags(): the parallelization flags of pw.x.

QuantumESPRESSOCommands.PhxConfig — Type

PhxConfig(; exe, chdir, options)

Create configurations for ph.x.

Arguments

exe::String="ph.x": the path to the executable.

chdir::Bool=true: whether to change directory to where the input file is stored when running ph.x. If false, stay in the current directory.

options::ParallelizationFlags=ParallelizationFlags(): the parallelization flags of ph.x.

QuantumESPRESSOCommands.Q2rxConfig — Type

Q2rxConfig(; exe, chdir, options)

Create configurations for q2r.x.

Arguments

exe::String="q2r.x": the path to the executable.

chdir::Bool=true: whether to change directory to where the input file is stored when running q2r.x. If false, stay in the current directory.

options::ParallelizationFlags=ParallelizationFlags(): the parallelization flags of q2r.x.

QuantumESPRESSOCommands.MatdynxConfig — Type

MatdynxConfig(; exe, chdir, options)

Create configurations for matdyn.x.

Arguments

exe::String="matdyn.x": the path to the executable.

chdir::Bool=true: whether to change directory to where the input file is stored when running matdyn.x. If false, stay in the current directory.

options::ParallelizationFlags=ParallelizationFlags(): the parallelization flags of matdyn.x.

QuantumESPRESSOCommands.DynmatxConfig — Type

DynmatxConfig(; exe, chdir, options)

Create configurations for dynmat.x.

Arguments

exe::String="dynmat.x": the path to the executable.

chdir::Bool=true: whether to change directory to where the input file is stored when running dynmat.x. If false, stay in the current directory.

options::ParallelizationFlags=ParallelizationFlags(): the parallelization flags of dynmat.x.

QuantumESPRESSOCommands.readconfig — Function

readconfig(cfgfile)

Read the configurations into an object from a file cfgfile.

QuantumESPRESSOCommands.makecmd — Function

makecmd(input; output, error, dir, use_script, mpi, main)

Make commands for QuantumESPRESSO executables.

Arguments

input: the path to the input file.

output=mktemp(parentdir(input))[1]: the path to the output file.

error=output: the path to the error file. By default, it logs into the output file.

dir=parentdir(input): change the working directory to dir. By default, it is the directory where the input file is stored.

use_script=false: if true, generate a shell script (with a random name) under dir and run it.

mpi=MpiexecConfig(): MPI configurations.

main: the configurations of the main executable.